@@ -9,31 +9,31 @@
   "author": "Vadim Matveev",
   "license": "MIT",
   "devDependencies": {
-    "@babel/core": "^7.15.8",
-    "@parcel/babel-preset-env": "^2.0.0",
-    "@parcel/packager-raw-url": "^2.0.0",
-    "@parcel/transformer-image": "^2.0.0",
-    "@parcel/transformer-stylus": "^2.0.0",
-    "@parcel/transformer-svg-react": "^2.0.0",
-    "@parcel/transformer-webmanifest": "^2.0.0",
+    "@babel/core": "^7.16.0",
+    "@parcel/babel-preset-env": "^2.0.1",
+    "@parcel/packager-raw-url": "^2.0.1",
+    "@parcel/transformer-image": "^2.0.1",
+    "@parcel/transformer-stylus": "^2.0.1",
+    "@parcel/transformer-svg-react": "^2.0.1",
+    "@parcel/transformer-webmanifest": "^2.0.1",
     "@svgr/parcel-plugin-svgr": "^5.5.0",
     "dotenv": "^10.0.0",
-    "eslint": "^8.1.0",
+    "eslint": "^8.2.0",
     "eslint-config-google": "^0.14.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-react": "^7.26.1",
-    "parcel": "^2.0.0",
+    "parcel": "^2.0.1",
     "prettier": "^2.4.1"
   },
   "dependencies": {
     "@babel/plugin-transform-runtime": "^7.16.0",
-    "browserslist": "^4.17.5",
+    "browserslist": "^4.17.6",
     "file-saver": "^2.0.5",
     "jszip": "^3.7.1",
     "react": "^17.0.2",
-    "react-bootstrap": "^2.0.1",
+    "react-bootstrap": "^2.0.2",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0",
+    "react-router-dom": "^6.0.1",
     "react-select": "^4.3.1",
     "react-string-replace": "^0.4.4",
     "react-swipeable": "^6.2.0",
